% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo_arm.R
\name{repo_arm}
\alias{repo_arm}
\alias{repo_arm_path}
\alias{repo_arm_type}
\alias{repo_arm_create}
\alias{repo_arm_insert}
\alias{repo_arm_contains}
\alias{repo_arm_remove}
\alias{repo_arm_find}
\alias{repo_arm_update}
\title{Internal repository arm API for developers}
\usage{
repo_arm(root, type, r_version)

repo_arm_path(arm, ...)

repo_arm_type(arm)

repo_arm_create(arm)

repo_arm_insert(arm, file)

repo_arm_contains(arm, file)

repo_arm_remove(arm, package, version)

repo_arm_find(arm, package, version)

repo_arm_update(arm)
}
\description{
Internal repository arm API for developers
}
\keyword{internal}
