% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_create}
\alias{repo_create}
\title{Create a CRAN-like package repository}
\usage{
repo_create(dir = ".", r_version = getRversion())
}
\arguments{
\item{dir}{Path to the directory to use as the repository root. Will be
created if it does not exist.}

\item{r_version}{The version of R to create binary distributions for.}
}
\value{
The path to the repository root directory, invisibly.
}
\description{
Create a CRAN-like package repository
}
