% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_insert}
\alias{repo_insert}
\title{Insert a package into a repository}
\usage{
repo_insert(repo, file, type, r_version = getRversion(), replace = FALSE)
}
\arguments{
\item{repo}{Path to the root directory of the package repository.}

\item{file}{Path to the source or binary package bundle to be added.}

\item{type}{Type of packages. One of: \code{"source"}, \code{"win.binary"}, \code{"mac.binary"}.}

\item{r_version}{For binary packages, the version of R they were built for.}

\item{replace}{Logical. Should the package be replaced if already present?}
}
\value{
Path(s) to the inserted package file(s), invisibly.
}
\description{
You can insert multiple packages with one call, as long as they are of the
same type and for the same version of R.
}
