% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_remove}
\alias{repo_remove}
\title{Remove a package from a repository}
\usage{
repo_remove(repo, package, version, type, r_version = getRversion())
}
\arguments{
\item{repo}{Path to the root directory of the package repository.}

\item{package}{The name of the package to remove.}

\item{version}{The version of the package to remove. Use \code{NULL} to remove all
versions.}

\item{type}{The type of the package to remove.}

\item{r_version}{If removing a binary package, the version of R that the
package was built for.}
}
\value{
The path(s) to the removed package file(s), invisibly.
}
\description{
Remove a package from a repository
}
