% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_update}
\alias{repo_update}
\title{Update the package index of a repository}
\usage{
repo_update(repo, type, r_version = getRversion())
}
\arguments{
\item{repo}{Path to the root directory of the package repository.}

\item{type}{Type of packages. One of: \code{"source"}, \code{"win.binary"}, \code{"mac.binary"}.}

\item{r_version}{For binary packages, the version of R they were built for.}
}
\description{
Do a full update of the package index in a repository. Functions in this
package update the index as they operate, so calling this should only be
necessary if your index has gone out of sync due to external changes.
}
\seealso{
Other functions to manage repositories: 
\code{\link{repo_create}()},
\code{\link{repo_insert}()},
\code{\link{repo_remove}()}
}
\concept{functions to manage repositories}
\concept{manage}
