% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cranrepo-package.R
\docType{package}
\name{cranrepo-package}
\alias{cranrepo}
\alias{cranrepo-package}
\title{cranrepo: Create and Maintain CRAN-like Repositories}
\description{
Basic tools for creating CRAN-like package repositories, adding and removing packages, and keeping the package index updated.
}
\author{
\strong{Maintainer}: Mikko Marttila \email{mikkmart@pm.me}

}
\keyword{internal}
