% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_create}
\alias{repo_create}
\title{Create a CRAN-like package repository}
\usage{
repo_create(root = ".", r_version = getRversion())
}
\arguments{
\item{root}{Path to the directory to use as the repository root.
Will be created if it does not exist.}

\item{r_version}{The version of R to create binary distribution trees for.}
}
\value{
Path to the repository root directory, invisibly.
}
\description{
Set up the directory structure and package indexes for
a CRAN-like package repository suitable to \code{\link[=install.packages]{install.packages()}} from.
}
\details{
Binary distribution trees are always created to avoid errors
when the repository is accessed via the \verb{file://} protocol.
}
\examples{
\dontrun{
# Create a package repository
repo_create("./repos/latest")
}
}
\seealso{
\code{\link[servr:httd]{servr::httd()}} to serve packages from the repository over HTTP.

Other functions to manage repositories: 
\code{\link{repo_insert}()},
\code{\link{repo_remove}()},
\code{\link{repo_update}()}
}
\concept{functions to manage repositories}
\concept{manage}
