% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_insert}
\alias{repo_insert}
\title{Insert a package into a repository}
\usage{
repo_insert(repo, file, type = NULL, r_version = getRversion())
}
\arguments{
\item{repo}{Path to the root directory of the package repository.}

\item{file}{Path to the source or binary package bundle to be added.}

\item{type}{The type of package to be added. Will be inferred from the file
name if left \code{NULL}.}

\item{r_version}{When inserting a binary package, the version of R that the
package was built for.}
}
\value{
The path to the inserted package file, invisibly.
}
\description{
Insert a package into a repository
}
