% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_remove}
\alias{repo_remove}
\title{Remove a package from a repository}
\usage{
repo_remove(
  repo,
  package,
  version,
  type,
  r_version = getRversion(),
  commit = FALSE
)
}
\arguments{
\item{repo}{Path to the root directory of the package repository.}

\item{package}{The name of the package to remove.}

\item{version}{The version of the package to remove.
Use \code{NULL} to remove all versions.}

\item{type}{Type of packages. One of: \code{"source"}, \code{"win.binary"}, \code{"mac.binary"}.}

\item{r_version}{For binary packages, the version of R they were built for.}

\item{commit}{Logical. If \code{FALSE} return the list of files that \emph{would be}
removed, without actually removing them.}
}
\value{
Path(s) to the removed package file(s), invisibly.
}
\description{
By default, shows a list of files that \emph{would be} removed.
Explicitly specify \code{commit = TRUE} to actually remove them.
}
