% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_serve}
\alias{repo_serve}
\title{Serve a repository over HTTP}
\usage{
repo_serve(repo, ...)
}
\arguments{
\item{repo}{Path to the root directory of the package repository.}

\item{...}{Additional arguments passed on to \code{\link[servr:httd]{servr::httd()}}.}
}
\description{
Serve a repository over HTTP
}
